Java Expense Manager: Java Swing Application for Track, Manage, Budget Expenses
This Java Swing application provides users with a convenient platform to track and manage their expenses efficiently. With easy input of expenses, the program calculates the total expenditure in real-time, offering valuable insights for budgeting and financial planning. It also notifies users if an expense exceeds the predefined limit. Whether for personal finance management or business expense tracking, this user-friendly solution helps users stay organized and in control of their finances.